+++ graphai/__init__.py
@@ -1,4 +1,5 @@
+from graphai.callback import Callback
 from graphai.graph import Graph
 from graphai.nodes import node, router
 
-__all__ = ["node", "router", "Graph"]
+__all__ = ["node", "router", "Callback", "Graph"]